requirements.txt
Flask==3.1.0
flask_cors==5.0.1
python-dotenv==1.1.0
google-generativeai==0.8.4
python-docx==1.1.2
lxml==5.3.1
requests==2.32.3
langchain-anthropic
langchain-core
llama_index
llama_index.readers.web